refactor: 메시지 및 조건들 상수화
