+++ hooks/src/lib/useCardHolderName/useCardHolderNameValidation.ts
@@ -1,6 +1,7 @@
 import { useState } from "react";
 
 import cardInputValidator from "../validators/cardInputValidator";
+import { VALIDATION_MESSAGES } from "../constants/cardCustomHook";
 
 const useCardHolderNameValidation = () => {
   const [errorState, setErrorState] = useState(false);
@@ -12,7 +13,7 @@ const useCardHolderNameValidation = () => {
       cardInputValidator.validateAlphabetInput(value);
 
     if (!isAlphabetInputValid) {
-      setErrorText("알파벳을 입력해 주세요.");
+      setErrorText(VALIDATION_MESSAGES.onlyEnglishAllowed);
       setErrorState(!isAlphabetInputValid);
 
       return false;
@@ -22,7 +23,7 @@ const useCardHolderNameValidation = () => {
       cardInputValidator.validateNumberInRange(value.length, 1, 15);
 
     if (!isValidCardHolderNameLength) {
-      setErrorText("15자 이내로 입력할 수 있습니다.");
+      setErrorText(VALIDATION_MESSAGES.invalidCardHolderName);
       setErrorState(!isValidCardHolderNameLength);
 
       return true;
