+++ hooks/src/lib/useCardPassword/useCardPasswordValidation.ts
@@ -1,6 +1,7 @@
 import { useState } from "react";
 
 import cardInputValidator from "../validators/cardInputValidator";
+import { VALIDATION_MESSAGES } from "../constants/cardCustomHook";
 
 const useCardPasswordValidation = () => {
   const [errorState, setErrorState] = useState(false);
@@ -11,7 +12,7 @@ const useCardPasswordValidation = () => {
     const isNumericInputValid = cardInputValidator.validateNumericInput(value);
 
     if (!isNumericInputValid) {
-      setErrorText("숫자를 입력해 주세요.");
+      setErrorText(VALIDATION_MESSAGES.onlyNumbersAllowed);
       setErrorState(!isNumericInputValid);
 
       return false;
@@ -23,7 +24,7 @@ const useCardPasswordValidation = () => {
     );
 
     if (!isValidCardPasswordLength) {
-      setErrorText("2자 이내로 입력할 수 있습니다.");
+      setErrorText(VALIDATION_MESSAGES.invalidCardPassword);
       setErrorState(!isValidCardPasswordLength);
 
       return true;
